Reuse SINGLE_CLUSTER feature Add cluster details page
