+++ libs/ui-lib/lib/ocm/components/clusterWizard/ClusterWizardContext.tsx
@@ -18,7 +18,6 @@ export type ClusterWizardContextType = {
   setWizardPerPage: (perPage: number) => void;
   updateUISettings: (data: UISettingsValues) => Promise<void>;
   uiSettings?: UISettingsValues;
-  isDisconnectedMode: boolean;
 };
 
 export const ClusterWizardContext = React.createContext<ClusterWizardContextType | null>(null);
