+++ libs/ui-lib/lib/ocm/components/clusterWizard/ClusterWizardContextProvider.tsx
@@ -81,12 +81,10 @@ const ClusterWizardContextProvider = ({
   cluster,
   infraEnv,
   permissions,
-  isDisconnectedMode = false,
 }: PropsWithChildren<{
   cluster?: Cluster;
   infraEnv?: InfraEnv;
   permissions?: AssistedInstallerOCMPermissionTypesListType;
-  isDisconnectedMode?: boolean;
 }>) => {
   const isSingleClusterFeatureEnabled = useFeature('ASSISTED_INSTALLER_SINGLE_CLUSTER_FEATURE');
   const [currentStepId, setCurrentStepId] = React.useState<ClusterWizardStepsType>();
@@ -250,7 +248,6 @@ const ClusterWizardContextProvider = ({
       setWizardPerPage,
       uiSettings,
       updateUISettings,
-      isDisconnectedMode,
     };
   }, [
     wizardStepIds,
@@ -262,7 +259,6 @@ const ClusterWizardContextProvider = ({
     clearAlerts,
     uiSettings,
     updateUISettings,
-    isDisconnectedMode,
   ]);
 
   if (!contextValue) {
