+++ libs/ui-lib/lib/ocm/hooks/index.ts
@@ -5,3 +5,4 @@ export { default as usePullSecret } from './usePullSecret';
 export { default as useClusterPreflightRequirements } from './useClusterPreflightRequirements';
 export { default as useUISettings } from './useUISettings';
 export { default as useInfraEnv } from './useInfraEnv';
+export { useFeatureDetection } from './use-feature-detection';
